Refactor theme into ThemeSwitcher
+ Fix gitignore to match lib for server only
