+++ .gitignore
@@ -14,8 +14,8 @@ dist/
 downloads/
 eggs/
 .eggs/
-lib/
-lib64/
+server/lib/
+server/lib64/
 parts/
 sdist/
 var/
