+++ client/src/routes/+layout.svelte
@@ -1,20 +1,6 @@
 <script lang="ts">
 	import '@picocss/pico';
-    import theme from '$lib/shared/stores/theme';
-	import { browser } from '$app/environment';
-
-	let isDarkMode = false;
-    theme.subscribe(t => {
-        isDarkMode = t === 'dark';
-    })
-    function handleTheme() {
-        const newTheme = isDarkMode ? 'light' : 'dark';
-        theme.set(newTheme);
-        document.getElementById('root')?.setAttribute('data-theme', newTheme);
-    };
-	if (browser) {
-		document.getElementById('root')?.setAttribute('data-theme', $theme);
-	}
+	import ThemeSwitcher from '$lib/components/ThemeSwitcher.svelte';
 </script>
 
 <header class="container-fluid">
@@ -27,14 +13,7 @@
 		</ul>
 		<ul>
 			<li>
-				<label
-					>Dark mode <input
-						type="checkbox"
-						role="switch"
-						bind:checked={isDarkMode}
-						on:click={handleTheme}
-					/></label
-				>
+				<ThemeSwitcher />
 			</li>
 		</ul>
 	</nav>
