+++ client/src/routes/+page.svelte
@@ -1,5 +1,5 @@
 <svelte:head>
-    <title>Home - Hako</title>
+	<title>Home - Hako</title>
 </svelte:head>
 <h1>Welcome to SvelteKit</h1>
 <p>Visit <a href="https://kit.svelte.dev">kit.svelte.dev</a> to read the documentation</p>
